IMPORTANT -- AARON: READ DESCRIPTION
I flipped the way the board is printed out onto the screen. location [0,0] is now at the bottom left. It just seemed logical that way.
no need to read further unless particularly curious.
adding better testing to inputMove() so accidental mistypes dont crash the program. We will probably be testing in this manner for a while.
all my half completed changes have been commented out so the game still runs.
and i disaprove of removing this.movePiece -- There is probably a good reason but i'm going to very dissapointed until i hear it..

Kudos for reading the whole message!!! -- Corbin